Optimize createListFromArrayLike() and Proxy's [[OwnPropertyKeys]] method https://bugs.webkit.org/show_bug.cgi?id=223928
Reviewed by Yusuke Suzuki.

JSTests:

* microbenchmarks/json-stringify-array-replacer.js:
Reduce running time from over 350ms to ~60ms.

* microbenchmarks/reflect-own-keys-proxy-2.js: Added.
* microbenchmarks/reflect-own-keys-proxy.js: Added.

Source/JavaScriptCore:

createListFromArrayLike() changes:

  1. Use toLength() / getIndex() methods that have fast paths.
  2. Remove RuntimeTypeMask and error messages from its signature: type checks are better
     performed in advance / inside a functor to keep the helper more versatile.
  3. Invert functor's return value to align with Structure::forEachProperty() and friends.
  4. Rename it to forEachInArrayLike() as no list is actually returned.

ProxyObject::performGetOwnPropertyNames() changes:

  1. Remove RuntimeTypeMask filtering as it's already performed by PropertyNameArray::add().
  2. Store target's keys in a HashSet for faster insertion / search.
  3. Don't populate `targetConfigurableKeys` for extensible target as it won't be used [1].
  4. Leverage return value of HashSet::remove() instead of using a helper.

This patch advances Proxy's [[OwnPropertyKeys]] microbenchmarks by 20-30%,
mainly due to createListFromArrayLike() changes. No behavior changes.

Also, utilizes forEachInArrayLike() for allow list of JSON.stringify().

[1]: https://tc39.es/ecma262/#sec-proxy-object-internal-methods-and-internal-slots-ownpropertykeys (step 20)

* runtime/JSONObject.cpp:
(JSC::Stringifier::Stringifier):
* runtime/JSObject.h:
(JSC::JSObject::getIndex const):
* runtime/JSObjectInlines.h:
(JSC::forEachInArrayLike):
(JSC::createListFromArrayLike): Deleted.
* runtime/ProxyObject.cpp:
(JSC::ProxyObject::performGetOwnPropertyNames):
* runtime/ReflectObject.cpp:
(JSC::JSC_DEFINE_HOST_FUNCTION):


Canonical link: https://commits.webkit.org/236034@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@275363 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/JavaScriptCore/runtime/JSObject.h
@@ -404,7 +404,7 @@ public:
         return JSValue();
     }
         
-    JSValue getIndex(JSGlobalObject* globalObject, unsigned i) const
+    JSValue getIndex(JSGlobalObject* globalObject, uint64_t i) const
     {
         if (JSValue result = tryGetIndexQuickly(i))
             return result;
